examples: add tcpsynbl.yaml for tracing the TCP SYN backlog
Background:

There are two types of backlogs when TCP connection happens:

- SYN Backlog, which buffers connections on receiving SYN packets
- Listen Backlog, which buffers connections after sockets are ESTABLISHED but before accepted by applications

In the world of web programming, it is often userful to monitor the size of backlogs,
and find out whether we need to increase the number of backlogs limit.

ebpf_exporter achieves that you can collect those metrics in a histogram and export,
and can nicely visualize in Grafana dashboard for example.