add support for fixed, user-defined buckets
Currently, ebpf_exporter supports exp2 and linear buckets. This is great
in many scenarios, but when the min bucket and max bucket are far apart,
it can result in a large number of total buckets which can overload
metric processing systems from a storage and cost perspective.

In some cases, it's preferable to have a small, fixed number of buckets
which can satisfy a use case while being cost-effective.

There are different ways this could be accomplished, but this commit
takes the approach of allowing the user to specify a fixed number of
pre-defined buckets, e.g.:

  - metrics:
      histograms:
      - bucket_type: fixed
        bucket_multiplier: 1e-06
        bucket_keys:
        - 100     # 100us
        - 1000    # 1ms
        - 250000  # 250ms
        - 1000000 # 1s
        - 2500000 # 2.5s
        - 5000000 # 5s

See examples/syslat.yaml for an example of how to set these buckets
within an eBPF program.