Publish binaries for x86_64 and aarch64
The binaries are built on Ubuntu 18.04 (Bionic), which comes with glibc 2.27.
This means that the resulting binary should work on the following distros
and their newer releases with newer glibc:

* Ubuntu Bionic (2.27)
* Debian Buster (2.28)
* CentOS 8 (2.28)

The resulting binary is dynamically linked against `libbcc.so`,
and it's the user's responsibility to supply that. On newer
Ubuntu and Debian releases it's in the `libbpfcc` package.