Pre-aggregate label sets to allow duplicate labels
When systemd restarts a service, it makes a fresh cgroup:

* https://github.com/systemd/systemd/issues/26101

For metrics that use `bpf_get_current_cgroup_id()` with a `cgroup` decoder
in a label that means that two different IDs would map to the same cgroup name.
As a result, prometheus is unhappy that two timeseries have the same set of
labels, breaking the exporter.

Let's pre-aggregate metrics to collapse these into the same label set
by summing up the values. In the cgroup example above we end up with just
one unique set of labels per service, encompassing events across restarts.

This assumes that all exported values are counters or do not have
conflicts as described above.

Let's also switch bpf map type for cgroup example to LRU hash, to account
for cgroup id churn, as we don't want to hit the map size limit over time.