ext4dist/xfsdist: fix the size of the latency hash map to be correct
ext4dist and xfsdist accumulate the latency histograms for multiple
operations, so the size of the latency hash maps must be the number of
latency slots multiplied by the number of operations. We define a new
F_MAX enum element and use it to size the BPF hash map.

This fixes issue #371 for ext4dist and xfsdist.