biolatency: fix the size of the latency hash map to be correct
This was previously sizing the BPF hash map used for the latency
histogram as the number of latency slots (plus one) times the maximum
number of disks. However, we accumulate latency histograms for
multiple operations, not a single one, so we need to also multiply the
hash map size by the number of expected operations in order to be
correct. Problems were typically hidden previously by the maximum
number of disks being large and the number of different types of
operations seen in practice typically being modest (five is typical on
our machines).

Much as the maximum number of disks is conservative, the maximum
number of different request operations is also set conservatively, to
the total of 14 REQ_OP_* things defined in include/linux/blk_types.h
as of 6.9-rc2.

This should correct issue #371 for biolatency.