Add padding label config key
Padding bytes in the map key struct need to be considered in the config
of the label. This is done by choosing a larger size for the label's value.
For most cases this works fine.

This commit adds a new config key for labels that allows to specify padding
bytes explicitly so they can be skipped by decoders.

Closes #369