examples/biolatency: correct the name of operation 35
Bio request operation 35 is REQ_OP_DRV_OUT, and should be called
'drv_out' in the label. Unfortunately when I added it in a previous
commit I duplicated the name of the previous entry so it was
mis-labeled as 'drv_in'. This is especially bad as it could lead to
label duplications in visible Prometheus metrics.