Add cgroup observer to prevent infinite fanotify growth
With fanotify we see every cgroup appear. Even if there are no metrics
associated with it and the cgroup quickly goes away, we keep the mapping
in memory. This mostly works fine, until you have an infinite churn of
cgroups, then memory starts growing indefinitely.

This commit adds a garbage collecting cache that would remove old entries
that were not requested in a long time.