netstacklat: first filter for sockets having a queue
It is more efficient to first filter out sockets with empty or small queues.
This is a tradeoff as our production use-case is to capture when the system
becomes overloaded.

name netstacklat_tcp_recv_timestamp
 - run_time_ns 15347880145 run_cnt 177786472 = 86.32 ns

name netstacklat_skb_consume_udp
 - run_time_ns 3096529442 run_cnt 33903931 = 91.33 ns

The performance gain is huge. Do remember that this is the average runtime cost
that is reduced, because we can skip recording many of these events.

Signed-off-by: Jesper Dangaard Brouer <hawk@kernel.org>