netstacklat: atomic counter for filter_nth_packet
We unfortunately need atomic counter update for the nth-filter.
This is because hooks like tcp-socket-read runs outside the socket
lock in a preempt/migrate-able user context.

We don't need accurate nth-counter across CPU, as this is
just a down-sampling mechanism.  Thus, we keep the PERCPU
array map and have nth-counter on a per CPU basis.  The
trick here is that in most cases the counter is only used
by the current running CPU, and the cache-line will mostly
be in a cache coherency Exclusive/Modified (MOESI) state,
which will cost less when doing atomic updates.

Manually testing on production showed 7ns runtime increase
(before 150.88 ns, after 157.67 ns).

Signed-off-by: Jesper Dangaard Brouer <hawk@kernel.org>